requirements.txt
certifi>=2018.11.29
chardet>=3.0.4
coreapi>=2.3.3
coreschema>=0.0.4
defusedxml>=0.5.0
Django>=2.1.5
django-allauth>=0.38.0
django-cors-headers>=2.4.0
django-rest-swagger>=2.2.0
django-webpack-loader>=0.6.0
djangorestframework>=3.9.1
httpie>=1.0.2
idna>=2.8
itypes>=1.1.0
Jinja2>=2.10
MarkupSafe>=1.1.0
mysqlclient==1.4.1
oauthlib>=3.0.1
openapi-codec>=1.3.2
pbr>=5.1.2
Pillow>=5.4.1
Pygments>=2.3.1
python3-openid>=3.1.0
pytz>=2018.9
requests>=2.21.0
requests-oauthlib>=1.2.0
simplejson>=3.16.0
six>=1.12.0
stevedore>=1.30.0
uritemplate>=3.0.0
urllib3>=1.24.1
virtualenv>=16.3.0
virtualenv-clone>=0.5.1
virtualenvwrapper>=4.8.2
